Extract a reusable method to send secure requests (csrf protected)
